Converting from raw socket to SslStream
Converted JamPackets to send and receive messages over SslStreams to allow for tsl security. JamServer, and JamClient also had to be modified to instanciate SslStreams and authenticate with certificates.

A self signed certifcate has been included in the solution directory for use when testing. This certificate does not contain any pertinant data and should not be used in release builds.