Server Side Messaging
Added server side message routing to JamServer. This required finishing up the Authentication code in the JamAccount class, as well as building data packing mechanisms for JamPackets and a Router/Interpreter combo for the server to receive and retransmit messages.

Still have to add similar logic to the client side.