Fixes to JamPacket Added Login to Example Client
Changed around a few thins in JamPacket to fix some issues with messaging. As strings were not being marshalled correctly, have changed message data to be serialized with json and utf8 encoding instead. JamPackets now receive the full message before exiting instead of just 1 byte... doh.

Added login functionality to login page in the example client, still a WIP need to add a page to navigate to inside of the example, and some code to log out again.