Preliminary Addition Client Disconnect Detection
Added some ground work to add in client disconnect detection. Reworked the packet router a little bit, and added an internal interpreter to the serverconnection object. Still need to add an internal interpreter to the client object. And will need to add in the polling function to determine when a user disconnects from the server, this will be in the form of a Ping message which as been added to the default data types. Polling will only be one way (server -> client), as client polling isn't really neccisary at this stage.