Refactored Message Interpreters to Events
Removed message interpreter interface, and refactored the way messages are handled.

Messages are now handled by Event Handlers. There is a Message Received event handler is the JamClient, and JamServer, and a Undelivered Message Received event handler in the server for handling messages addressed to users who are not online.

Event Handlers will be used in the future, next event handlers will be for server connections and disconnections.