Added Account Registration
Added "Register" button to the login page of the example client, this new button sends a message requesting an account be created rather than attempting to log into one. This is the first message that is managed only by the Example Client/Server and is outside of the JamLib class library.