User List Population
Users list in example client is now populated by the example server, and is updated in real time when users come online and go offline.

Still need to add some sorting to the list to bubble online users to the top, and change the forecolour of the online users to make it more identifieable.