Added Unread Message Count
Moved unseen messages getter partially into the displayable account, and made displayable account implement ipropertychanged, this allows for each account list item to display the number of unread messages it contains.

The number of unread messages is displayed in a small red bubble for each account, and the sum of these totals is shown under the logged in user account at the top.