Moved Login Response to JamClient
Moved the handling of the Login Respons message out of the reponsibility of the implementer and into the internal client interpreter of the JamClient. A new event has been added for when a login result is received, and this is used in place of the message interpreters self handling code.